realm_gc/rgc_control/src/cmarl/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
colour==0.1.5
contourpy==1.3.0
control==0.10.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.8.0
equinox==0.11.6
etils==1.9.4
exceptiongroup==1.2.2
executing==2.1.0
flax==0.8.4
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
humanize==4.10.0
idna==3.9
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.27.0
jax==0.4.26
jax-cuda12-pjrt==0.4.26
jax-cuda12-plugin==0.4.26
jaxlib==0.4.26
jaxtyping==0.2.34
jedi==0.19.1
jraph==0.0.6.dev0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.1.4
nvidia-cuda-cupti-cu12==12.6.68
nvidia-cuda-nvcc-cu12==12.6.68
nvidia-cuda-nvrtc-cu12==12.6.68
nvidia-cuda-runtime-cu12==12.6.68
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.6.59
nvidia-cusolver-cu12==11.6.4.69
nvidia-cusparse-cu12==12.5.3.3
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.6.68
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.6.3
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
tensorflow-probability==0.24.0
tensorstore==0.1.65
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wandb==0.18.0
wcwidth==0.2.13
zipp==3.20.2
